cloudapi: specify min_size type
In the filesystem object, use x-go-type to specify
uint64 as the type for the min_size field. Avoids
a type conversion in v2.go.

+++ internal/cloudapi/v2/handler.go
@@ -187,7 +187,7 @@ func (h *apiHandlers) PostCompose(ctx echo.Context) error {
 			fsCustomizations = append(fsCustomizations,
 				blueprint.FilesystemCustomization{
 					Mountpoint: f.Mountpoint,
-					MinSize:    uint64(f.MinSize),
+					MinSize:    f.MinSize,
 				},
 			)
 		}
